+++ src/CleanRegex/Internal/Match/GroupBy/TextsStrategy.php
@@ -8,10 +8,11 @@ class TextsStrategy implements Strategy
 {
     public function transform(array $groups, RawMatchesOffset $matches): array
     {
+        $closure = static function (IRawMatch $match) {
+            return $match->getText();
+        };
         foreach ($groups as &$group) {
-            $group = \array_map(static function (IRawMatch $match) {
-                return $match->getText();
-            }, $group);
+            $group = \array_map($closure, $group);
         }
         return $groups;
     }